ItemDB integrated with ItemManager
