--- Assets/Scripts/ItemDB.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 43d39bd0278e941138028e04c2b91803
-timeCreated: 1442404615
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
