+++ Assets/prefab/Inventory.prefab.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a47a5b7b3bbd34353bd07182f90530ca
+timeCreated: 1442567202
+licenseType: Free
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
