+++ Assets/scene/UITest.unity
@@ -85,6 +85,48 @@ NavMeshSettings:
     cellSize: .166666672
     manualCellSize: 0
   m_NavMeshData: {fileID: 0}
+--- !u!1001 &69461903
+Prefab:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    m_TransformParent: {fileID: 0}
+    m_Modifications:
+    - target: {fileID: 423722, guid: a47a5b7b3bbd34353bd07182f90530ca, type: 2}
+      propertyPath: m_LocalPosition.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 423722, guid: a47a5b7b3bbd34353bd07182f90530ca, type: 2}
+      propertyPath: m_LocalPosition.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 423722, guid: a47a5b7b3bbd34353bd07182f90530ca, type: 2}
+      propertyPath: m_LocalPosition.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 423722, guid: a47a5b7b3bbd34353bd07182f90530ca, type: 2}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 423722, guid: a47a5b7b3bbd34353bd07182f90530ca, type: 2}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 423722, guid: a47a5b7b3bbd34353bd07182f90530ca, type: 2}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 423722, guid: a47a5b7b3bbd34353bd07182f90530ca, type: 2}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 423722, guid: a47a5b7b3bbd34353bd07182f90530ca, type: 2}
+      propertyPath: m_RootOrder
+      value: 7
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+  m_ParentPrefab: {fileID: 100100000, guid: a47a5b7b3bbd34353bd07182f90530ca, type: 2}
+  m_IsPrefabParent: 0
 --- !u!1 &450001411
 GameObject:
   m_ObjectHideFlags: 0
